+++ queuing-system/queuing-system-for-java/src/main/java/io/chan/queuingsystemforjava/domain/seat/dto/response/SeatGradeElement.java
@@ -2,10 +2,12 @@ package io.chan.queuingsystemforjava.domain.seat.dto.response;
 
 import io.chan.queuingsystemforjava.domain.seat.SeatGrade;
 
+import java.math.BigDecimal;
+
 public record SeatGradeElement (
         Long seatGradeId,
         String gradeName,
-        Long price
+        BigDecimal price
 ){
     public static SeatGradeElement of(SeatGrade seatGrade) {
         return new SeatGradeElement(seatGrade.getSeatGradeId(), seatGrade.getGradeName(), seatGrade.getPrice());
